Make it actually work for longs URLs.
Gin doesn't seem to have a way of ignoring forward-slashes, so giving a link with them would end up getting you at the wrong router and wouldn't work. So let's use a hack and replace all forward-slashes to | which are not meant to be used in links and when we receive the link on the router, replace all the | back to forward-slashes. Maybe I'll find a better way, but this does work.