Better handle URLs by using base64URL encoded strings.
Previously, the way it was done was replacing characters that generated
issues when sending from browser to server. Now, the instead of sending
raw URLs with replaced characters, they're encoded to base64URL strings,
which are then decoded when received.

The hash that the user receives is also encoded now.